Add two new models: - DirGroupedModel and DirGroupedProxyModel
DirGroupedModel is an administrative model. It mainstains which groups are created and created models for each group.

DirGroupedProxyModel contains the filter to show the content belonging to that group. It is also going to have sorting capabilities which means that (technically) every individual group can be sorted how you want it.

Displaying the above models in QML is not possible yet. You can display every individual model with some tricks, but there is no way to get a nice group overview yet.